requirements.txt
paho-mqtt
Flask
pymysql
configparser